Add sd-notify support
If suricata is being run under systemd, this allows services which
depend on it to start only after it is up and running.

This helps if, say, said services want to be able to connect to suricata
control socket or some such, without having to timeout and retry in a
loop or resorting to other, more complicated, schemes.